refactor: move nats domain structs to `domain` package
